Move llm_token to cloud_api_client
