+++ crates/client/src/llm_token.rs
@@ -4,7 +4,7 @@ use cloud_llm_client::{EXPIRED_LLM_TOKEN_HEADER_NAME, OUTDATED_LLM_TOKEN_HEADER_
 use gpui::{
     App, AppContext as _, Context, Entity, EventEmitter, Global, ReadGlobal as _, Subscription,
 };
-use language_model::LlmApiToken;
+use cloud_api_client::LlmApiToken;
 use std::sync::Arc;
 
 pub trait NeedsLlmTokenRefresh {
