+++ crates/language_model/Cargo.toml
@@ -19,8 +19,6 @@ test-support = []
 anyhow.workspace = true
 credentials_provider.workspace = true
 base64.workspace = true
-cloud_api_client.workspace = true
-cloud_api_types.workspace = true
 collections.workspace = true
 env_var.workspace = true
 futures.workspace = true
@@ -33,7 +31,6 @@ log.workspace = true
 parking_lot.workspace = true
 serde.workspace = true
 serde_json.workspace = true
-smol.workspace = true
 thiserror.workspace = true
 util.workspace = true
