+++ crates/language_models/Cargo.toml
@@ -21,6 +21,7 @@ aws_http_client.workspace = true
 base64.workspace = true
 bedrock = { workspace = true, features = ["schemars"] }
 client.workspace = true
+cloud_api_client.workspace = true
 cloud_api_types.workspace = true
 collections.workspace = true
 component.workspace = true
